fix: resolve hydration mismatches, add document delete, fix bigint counts
✏️ Changed:
- 🐛 Fix hydration mismatch across 22 pages using useMounted() guard for React Query loading states
- 🐛 Fix backoffice tenants correlated subquery using raw SQL column references instead of Drizzle interpolation
- 🐛 Add ::int cast to all COUNT(*) SQL queries to prevent bigint-to-string serialization
- 🗑️ Add DELETE handler for documents API (OCR_PROCESSING/DRAFT only) with Supabase Storage file cleanup
- 🎨 Add delete button with confirmation modal to document side panel

➕ Added:
- 🏗️ src/lib/hooks/use-mounted.ts — shared hydration-safe hook for SSR/client render consistency
- 🔧 deleteDoc mutation in useDocumentMutations hook

+++ src/app/api/backoffice/rules/route.ts
@@ -77,10 +77,10 @@ export async function GET(request: NextRequest) {
     db
       .select({
         total: count(),
-        graduated: sql<number>`COUNT(*) FILTER (WHERE ${aiExtractionRules.isGraduated} = true)`,
-        prompt: sql<number>`COUNT(*) FILTER (WHERE ${aiExtractionRules.isGraduated} = false AND ${aiExtractionRules.confidence} >= '0.50')`,
-        low: sql<number>`COUNT(*) FILTER (WHERE ${aiExtractionRules.confidence} < '0.50')`,
-        tenantCount: sql<number>`COUNT(DISTINCT ${aiExtractionRules.tenantId})`,
+        graduated: sql<number>`COUNT(*) FILTER (WHERE ${aiExtractionRules.isGraduated} = true)::int`,
+        prompt: sql<number>`COUNT(*) FILTER (WHERE ${aiExtractionRules.isGraduated} = false AND ${aiExtractionRules.confidence} >= '0.50')::int`,
+        low: sql<number>`COUNT(*) FILTER (WHERE ${aiExtractionRules.confidence} < '0.50')::int`,
+        tenantCount: sql<number>`COUNT(DISTINCT ${aiExtractionRules.tenantId})::int`,
       })
       .from(aiExtractionRules),
   ]);
